# Use alternate runtime APIs to reduce allocs / cpu