```
benchmark                               old ns/op     new ns/op     delta
BenchmarkLocal_StartSpanFromContext-8   1382          718           -48.05%

benchmark                               old allocs    new allocs    delta
BenchmarkLocal_StartSpanFromContext-8   6             4             -33.33%

benchmark                               old bytes     new bytes     delta
BenchmarkLocal_StartSpanFromContext-8   376           224           -40.43%
```