@jacobmarble 2019-03-08
Added a small commit to the PR, to help understand performance composition. Thanks for this Stuart!