@stuartcarnie 2019-03-08
Good call - allows us to understand the relative perf of those runtime functions